Restructure how MDX shortcodes are imported
