+++ .eslintignore
@@ -1,4 +1,3 @@
 gatsby/
 gatsby*
-mdxImportGen.js
 public
