--- src/mdxGlobalComponents.js
@@ -1,193 +1,0 @@
-// AUTO GENERATED FILE
-
-import { Endpoint } from './components/APIDocs/Endpoint'
-import { MethodTags } from './components/APIDocs/MethodTags'
-import { Accordion } from './components/Accordion'
-import { AnchorScrollNavbar } from './components/AnchorScrollNavbar'
-import { AnimateIntoView } from './components/AnimateIntoView'
-import { AnimatedBurger } from './components/AnimatedBurger'
-import { ArrayCTA } from './components/ArrayCTA'
-import { BasicHedgehogImage } from './components/BasicHedgehogImage'
-import { BlockQuote } from './components/BlockQuote'
-import { Blog } from './components/Blog'
-import { BlogFooter } from './components/BlogFooter'
-import { BorderWrapper } from './components/BorderWrapper'
-import { Breadcrumbs } from './components/Breadcrumbs'
-import { CallToAction } from './components/CallToAction'
-import { Benefits } from './components/Careers/Benefits'
-import { CareersHero } from './components/Careers/CareersHero'
-import { InterviewProcess } from './components/Careers/InterviewProcess'
-import { OpenRoles } from './components/Careers/OpenRoles'
-import { Transparency } from './components/Careers/Transparency'
-import { WhyWereHere } from './components/Careers/WhyWereHere'
-import { WorkingAtPostHog } from './components/Careers/WorkingAtPostHog'
-import { Chip } from './components/Chip'
-import { CodeBlock } from './components/CodeBlock'
-import { CompensationCalculator } from './components/CompensationCalculator'
-import { Contact } from './components/Contact'
-import { ContactForm } from './components/ContactForm'
-import { Container } from './components/Container'
-import { ContributorCard } from './components/ContributorCard'
-import { ContributorSearch } from './components/ContributorSearch'
-import { ContributorsChart } from './components/ContributorsChart'
-import { Customers } from './components/Customers'
-import { DarkModeToggle } from './components/DarkModeToggle'
-import { DemoScheduler } from './components/DemoScheduler'
-import { DeployOption } from './components/DeployOption'
-import { DocsPageSurvey } from './components/DocsPageSurvey'
-import { DocsSearch } from './components/DocsSearch'
-import { FeaturesComparisonTable } from './components/FeaturesComparisonTable'
-import { FeaturesNav } from './components/FeaturesNav'
-import { FeaturesSelfHost } from './components/FeaturesSelfHost'
-import { FloatedImage } from './components/FloatedImage'
-import { Footer } from './components/Footer'
-import { GetStartedModal } from './components/GetStartedModal'
-import { GithubIcon } from './components/GithubIcon'
-import { Heading } from './components/Heading'
-import { HiddenSection } from './components/HiddenSection'
-import { Home } from './components/Home'
-import { HostingOption } from './components/HostingOption'
-import { ImageBlock } from './components/ImageBlock'
-import { InlineCode } from './components/InlineCode'
-import { LandingPageCallToAction } from './components/LandingPage/LandingPageCallToAction'
-import { LibraryStats } from './components/LibraryStats'
-import { Link } from './components/Link'
-import { Logo } from './components/Logo'
-import { MainNav } from './components/MainNav'
-import { MdxAnchorHeaders } from './components/MdxAnchorHeaders'
-import { Modal } from './components/Modal'
-import { NewsletterForm } from './components/NewsletterForm'
-import { NotFoundPage } from './components/NotFoundPage'
-import { OtherFeaturesBlock } from './components/OtherFeaturesBlock'
-import { OverflowXSection } from './components/OverflowXSection'
-import { PageHeader } from './components/PageHeader'
-import { PlanComparisonTable } from './components/PlanComparisonTable'
-import { PostCard } from './components/PostCard'
-import { CloudVsSelfHost } from './components/Pricing/CloudVsSelfHost'
-import { FAQs } from './components/Pricing/FAQs'
-import { PlanComparison } from './components/Pricing/PlanComparison'
-import { PricingHero } from './components/Pricing/PricingHero'
-import { PricingSlider } from './components/Pricing/PricingSlider'
-import { PricingTable } from './components/Pricing/PricingTable'
-import { Quote } from './components/Pricing/Quote'
-import { Savings } from './components/Pricing/Savings'
-import { ProductAnchorNavbar } from './components/ProductAnchorNavbar'
-import { ProductFeature } from './components/ProductFeature'
-import { ProductFeaturePlugin } from './components/ProductFeaturePlugin'
-import { ProductFooter } from './components/ProductFooter'
-import { ProductSectionHeader } from './components/ProductSectionHeader'
-import { ResponsiveAnchor } from './components/ResponsiveAnchor'
-import { RocketHorizontalRule } from './components/RocketHorizontalRule'
-import { DesignedForYourStackBlock } from './components/Sections/DesignedForYourStackBlock'
-import { FeaturedSectionTextLeft } from './components/Sections/FeaturedSectionTextLeft'
-import { FeaturedSectionTextRight } from './components/Sections/FeaturedSectionTextRight'
-import { FeaturedSectionTripleImage } from './components/Sections/FeaturedSectionTripleImage'
-import { SignupModal } from './components/SignupModal'
-import { SliderNav } from './components/SliderNav'
-import { Spacer } from './components/Spacer'
-import { StarRepoButton } from './components/StarRepoButton'
-import { StartNowButton } from './components/StartNowButton'
-import { Structure } from './components/Structure'
-import { TableOfContents } from './components/TableOfContents'
-import { TeamQuote } from './components/TeamQuote'
-import { Tooltip } from './components/Tooltip'
-import { VisitLibrary } from './components/VisitLibrary'
-import { WorkableSnippet } from './components/WorkableSnippet'
-import { ZoomImage } from './components/ZoomImage'
-
-export const shortcodes = {
-    Endpoint,
-    MethodTags,
-    Accordion,
-    AnchorScrollNavbar,
-    AnimateIntoView,
-    AnimatedBurger,
-    ArrayCTA,
-    BasicHedgehogImage,
-    BlockQuote,
-    Blog,
-    BlogFooter,
-    BorderWrapper,
-    Breadcrumbs,
-    CallToAction,
-    Benefits,
-    CareersHero,
-    InterviewProcess,
-    OpenRoles,
-    Transparency,
-    WhyWereHere,
-    WorkingAtPostHog,
-    Chip,
-    CodeBlock,
-    CompensationCalculator,
-    Contact,
-    ContactForm,
-    Container,
-    ContributorCard,
-    ContributorSearch,
-    ContributorsChart,
-    Customers,
-    DarkModeToggle,
-    DemoScheduler,
-    DeployOption,
-    DocsPageSurvey,
-    DocsSearch,
-    FeaturesComparisonTable,
-    FeaturesNav,
-    FeaturesSelfHost,
-    FloatedImage,
-    Footer,
-    GetStartedModal,
-    GithubIcon,
-    Heading,
-    HiddenSection,
-    Home,
-    HostingOption,
-    ImageBlock,
-    InlineCode,
-    LandingPageCallToAction,
-    LibraryStats,
-    Link,
-    Logo,
-    MainNav,
-    MdxAnchorHeaders,
-    Modal,
-    NewsletterForm,
-    NotFoundPage,
-    OtherFeaturesBlock,
-    OverflowXSection,
-    PageHeader,
-    PlanComparisonTable,
-    PostCard,
-    CloudVsSelfHost,
-    FAQs,
-    PlanComparison,
-    PricingHero,
-    PricingSlider,
-    PricingTable,
-    Quote,
-    Savings,
-    ProductAnchorNavbar,
-    ProductFeature,
-    ProductFeaturePlugin,
-    ProductFooter,
-    ProductSectionHeader,
-    ResponsiveAnchor,
-    RocketHorizontalRule,
-    DesignedForYourStackBlock,
-    FeaturedSectionTextLeft,
-    FeaturedSectionTextRight,
-    FeaturedSectionTripleImage,
-    SignupModal,
-    SliderNav,
-    Spacer,
-    StarRepoButton,
-    StartNowButton,
-    Structure,
-    TableOfContents,
-    TeamQuote,
-    Tooltip,
-    VisitLibrary,
-    WorkableSnippet,
-    ZoomImage,
-}
