+++ src/templates/Plain.js
@@ -1,30 +1,15 @@
-import React from 'react'
-import Layout from 'components/Layout'
-import { MDXRenderer } from 'gatsby-plugin-mdx'
 import { MDXProvider } from '@mdx-js/react'
+import Layout from 'components/Layout'
 import { SEO } from 'components/seo'
-import { CodeBlock } from '../components/CodeBlock'
-import { shortcodes } from '../mdxGlobalComponents'
-import { H1, H2, H3, H4, H5, H6 } from 'components/MdxAnchorHeaders'
-import Link from 'components/Link'
-
-const A = (props) => <Link {...props} className="text-red hover:text-red font-semibold" />
+import { MDXRenderer } from 'gatsby-plugin-mdx'
+import React from 'react'
+import { shortcodes } from '../shortcodes'
 
 export default function Plain({ data }) {
     const { pageData } = data
     const { body, excerpt } = pageData
     const { title, featuredImage, description, showTitle } = pageData?.frontmatter
-    const components = {
-        h1: H1,
-        h2: H2,
-        h3: H3,
-        h4: H4,
-        h5: H5,
-        h6: H6,
-        pre: CodeBlock,
-        a: A,
-        ...shortcodes,
-    }
+
     return (
         <Layout>
             <SEO
@@ -35,7 +20,7 @@ export default function Plain({ data }) {
             />
             <article className="max-w-2xl mx-auto my-12 md:my-24 px-4 article-content">
                 {showTitle && <h1 className="text-center">{title}</h1>}
-                <MDXProvider components={components}>
+                <MDXProvider components={shortcodes}>
                     <MDXRenderer>{body}</MDXRenderer>
                 </MDXProvider>
             </article>
